Error handling for new preview code
The old more complicated solution checking error codes is no longer
needed, as any error is now printed from stdout to the tab. We're not
dependent on Fuse being able to send us the error over the Fuse
protocol.

+++ building.py
@@ -7,29 +7,36 @@ from .fuse_util import getFusePathFromSettings
 from .log import log
 
 class BuildInstance(threading.Thread):
-	def __init__(self, cmd, title):
+	def __init__(self, cmd, title, fuseNotFoundHandler):
 		threading.Thread.__init__(self)
 		self.cmd = cmd
 		self.daemon = True
 		self.output = OutputView(title)
+		self.fuseNotFoundHandler = fuseNotFoundHandler
+		self.process = None
 
 	def run(self):
 		log().info("-- Opening subprocess %s", str(self.cmd))
-		#TODO handle exceptions and errors
-		if os.name == "nt":
-			CREATE_NO_WINDOW = 0x08000000
-			self.process = subprocess.Popen(self.cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, creationflags=CREATE_NO_WINDOW)
-		else:
-			self.process = subprocess.Popen(self.cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+		try:
+			if os.name == "nt":
+				CREATE_NO_WINDOW = 0x08000000
+				self.process = subprocess.Popen(self.cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, creationflags=CREATE_NO_WINDOW)
+			else:
+				self.process = subprocess.Popen(self.cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+		except:
+			self.fuseNotFoundHandler()
+			self.output.close()
+			return
 		for line in iter(self.process.stdout.readline,b''):
 			self.output.append(line.decode("utf-8"))
 		self.process.wait()
 
 	def stop(self):
-		try:
-			self.process.kill()
-		except ProcessLookupError:
-			pass #It died by itself, which is fine
+		if self.process:
+			try:
+				self.process.kill()
+			except ProcessLookupError:
+				pass #It died by itself, which is fine
 		self.output.close()
 
 class OutputView:
@@ -44,13 +51,17 @@ class OutputView:
 		self.view.run_command("append", {"characters": line})
 
 	def close(self):
-		window = self.view.window()
-		groupIndex, viewIndex = window.get_view_index(self.view)
-		window.run_command("close_by_index", { "group": groupIndex, "index": viewIndex })
+		try:
+			window = self.view.window()
+			groupIndex, viewIndex = window.get_view_index(self.view)
+			window.run_command("close_by_index", { "group": groupIndex, "index": viewIndex })
+		except:
+			pass #Failing to close a tab is not critical
 
 class BuildManager:
-	def __init__(self):
+	def __init__(self, fuseNotFoundHandler):
 		self.builds = {}
+		self.fuseNotFoundHandler = fuseNotFoundHandler
 
 	def preview(self, target, path):
 		fusePath = getFusePathFromSettings()
@@ -64,6 +75,6 @@ class BuildManager:
 	def __start(self, target, cmd, name):
 		if name in self.builds:
 			self.builds[name].stop()
-		build = BuildInstance(cmd, name)
+		build = BuildInstance(cmd, name, self.fuseNotFoundHandler)
 		self.builds[name] = build
 		build.start()
